Split each marin3r subcommand into its own file for clarity
