+++ cmd/webhook.go
@@ -1,0 +1,131 @@
+/*
+
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cmd
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+	apimachineryruntime "k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/healthz"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	"sigs.k8s.io/controller-runtime/pkg/manager/signals"
+	"sigs.k8s.io/controller-runtime/pkg/webhook"
+
+	marin3rv1alpha1 "github.com/3scale-ops/marin3r/apis/marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale-ops/marin3r/apis/operator.marin3r/v1alpha1"
+	"github.com/3scale-ops/marin3r/pkg/webhooks/podv1mutator"
+	// +kubebuilder:scaffold:imports
+)
+
+var (
+	webhookPort        int
+	webhookTLSCertDir  string
+	webhookTLSKeyName  string
+	webhookTLSCertName string
+)
+
+var (
+	// Webhook subcommand
+	webhookCmd = &cobra.Command{
+		Use:   "webhook",
+		Short: "Run the webhook server",
+		Run:   runWebhook,
+	}
+)
+
+var (
+	webhookScheme = apimachineryruntime.NewScheme()
+)
+
+func init() {
+	utilruntime.Must(clientgoscheme.AddToScheme(webhookScheme))
+	utilruntime.Must(operatorv1alpha1.AddToScheme(webhookScheme))
+	utilruntime.Must(marin3rv1alpha1.AddToScheme(webhookScheme))
+	utilruntime.Must(operatorv1alpha1.AddToScheme(webhookScheme))
+	// +kubebuilder:scaffold:scheme
+
+	rootCmd.AddCommand(webhookCmd)
+
+	// Webhook flags
+	webhookCmd.Flags().IntVar(&webhookPort, "webhook-port", int(operatorv1alpha1.DefaultWebhookPort), "The port where the pod mutator webhook server will listen.")
+	webhookCmd.Flags().StringVar(&webhookTLSCertDir, "tls-dir", "/apiserver.local.config/certificates", "The path where the certificate and key for the webhook are located.")
+	webhookCmd.Flags().StringVar(&webhookTLSCertName, "tls-cert-name", "apiserver.crt", "The file name of the certificate for the webhook.")
+	webhookCmd.Flags().StringVar(&webhookTLSKeyName, "tls-key-name", "apiserver.key", "The file name of the private key for the webhook.")
+}
+
+func runWebhook(cmd *cobra.Command, args []string) {
+
+	ctrl.SetLogger(zap.New(zap.UseDevMode(debug)))
+	printVersion()
+
+	cfg := ctrl.GetConfigOrDie()
+
+	mgr, err := ctrl.NewManager(cfg, ctrl.Options{
+		Scheme:                 webhookScheme,
+		MetricsBindAddress:     "0",
+		HealthProbeBindAddress: probeAddr,
+		Port:                   webhookPort,
+		LeaderElection:         false,
+	})
+	if err != nil {
+		setupLog.Error(err, "unable to start manager")
+		os.Exit(1)
+	}
+
+	// Setup the webhook
+	hookServer := mgr.GetWebhookServer()
+	hookServer.CertDir = webhookTLSCertDir
+	hookServer.KeyName = webhookTLSKeyName
+	hookServer.CertName = webhookTLSCertName
+	hookServer.Port = webhookPort
+
+	// Register the Pod mutating webhook
+	ctrl.Log.Info("registering the pod mutating webhook with webhook server")
+	hookServer.Register(podv1mutator.MutatePath, &webhook.Admission{Handler: &podv1mutator.PodMutator{Client: mgr.GetClient()}})
+
+	// Register the EnvoyConfig validating webhook
+	if err = (&marin3rv1alpha1.EnvoyConfig{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "EnvoyConfig")
+		os.Exit(1)
+	}
+
+	// Register the EnvoyDeployment validating webhook
+	if err = (&operatorv1alpha1.EnvoyDeployment{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "EnvoyDeployment")
+		os.Exit(1)
+	}
+
+	if err := mgr.AddHealthzCheck("health", healthz.Ping); err != nil {
+		setupLog.Error(err, "unable to set up health check")
+		os.Exit(1)
+	}
+	if err := mgr.AddReadyzCheck("check", healthz.Ping); err != nil {
+		setupLog.Error(err, "unable to set up ready check")
+		os.Exit(1)
+	}
+
+	setupLog.Info("starting the webhook")
+	if err := mgr.Start(signals.SetupSignalHandler()); err != nil {
+		setupLog.Error(err, "controller manager exited non-zero")
+		os.Exit(1)
+	}
+}
